+++ elide-integration-tests/src/test/groovy/com/yahoo/elide/tests/GraphQLIT.groovy
@@ -7,6 +7,7 @@ package com.yahoo.elide.tests
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.node.ArrayNode
 import com.fasterxml.jackson.databind.node.JsonNodeFactory
 import com.fasterxml.jackson.databind.node.ObjectNode
 import com.jayway.restassured.RestAssured
@@ -152,6 +153,39 @@ class GraphQLIT extends AbstractIntegrationTestInitializer {
         runQueryWithExpectedResult(graphQLQuery, expectedResponse)
     }
 
+    @Test(priority = 5)
+    void runUpdateAndFetchDifferentTransactionsBatch() {
+        def graphQLQuery = """
+        mutation {
+          book(op: UPSERT, data:{id:"abc", title: "my book created in batch!"}) {
+            edges {
+              node {
+                id
+                title
+              }
+            }
+          }
+        }
+        """
+        def graphQLQuery2 = """
+        query {
+          book(ids: "3") {
+            edges {
+              node {
+                id
+                title
+              }
+            }
+          }
+        }
+        """
+        def expectedResponse = """
+        [{"data":{"book":{"edges":[{"node":{"id":"3","title":"my book created in batch!"}}]}}},{"data":{"book":{"edges":[{"node":{"id":"3","title":"my book created in batch!"}}]}}}]
+        """
+
+        compareJsonObject(runQuery(toJsonArray(toJsonNode(graphQLQuery), toJsonNode(graphQLQuery2))), expectedResponse)
+    }
+
     private void runQueryWithExpectedResult(String graphQLQuery, Map<String, Object> variables, String expected) {
         compareJsonObject(runQuery(graphQLQuery, variables), expected)
     }
@@ -167,21 +201,41 @@ class GraphQLIT extends AbstractIntegrationTestInitializer {
     }
 
     private ValidatableResponse runQuery(String query, Map<String, Object> variables) {
+        return runQuery(toJsonQuery(query, variables))
+    }
+
+    private ValidatableResponse runQuery(String query) {
         return RestAssured.given()
                 .contentType(MediaType.APPLICATION_JSON)
                 .accept(MediaType.APPLICATION_JSON)
-                .body(toJsonQuery(query, variables))
+                .body(query)
                 .post("/graphQL")
                 .then()
                 .statusCode(HttpStatus.SC_OK)
     }
 
+    private String toJsonArray(JsonNode... nodes) {
+        ArrayNode arrayNode = JsonNodeFactory.instance.arrayNode()
+        for(JsonNode node : nodes) {
+            arrayNode.add(node)
+        }
+        return objectMapper.writeValueAsString(arrayNode)
+    }
+
     private String toJsonQuery(String query, Map<String, Object> variables) {
+        return objectMapper.writeValueAsString(toJsonNode(query, variables))
+    }
+
+    private JsonNode toJsonNode(String query) {
+        return toJsonNode(query, null)
+    }
+
+    private JsonNode toJsonNode(String query, Map<String, Object> variables) {
         ObjectNode graphqlNode = JsonNodeFactory.instance.objectNode()
         graphqlNode.put("query", query)
         if (variables != null) {
             graphqlNode.set("variables", objectMapper.valueToTree(variables))
         }
-        return objectMapper.writeValueAsString(graphqlNode)
+        return graphqlNode
     }
 }